# Use WithLabelValues instead of With for metrics in distributed.go