The main goal of this is to reduce the CPU usage of the cache proxy.

`metric.WithLabelValues` is roughly 3 times faster than `metric.With`, partially because the latter requires creating a map and assigning values to it.

Benchmark results:
```
goos: linux
goarch: amd64
cpu: AMD Ryzen 9 7900X 12-Core Processor
                                          │    base     │            findmissing2             │
                                          │   sec/op    │   sec/op     vs base                │
Read/LookasideDistPebble10-24               850.5n ± 2%   652.4n ± 2%  -23.29% (p=0.000 n=11)
Read/LookasideDistPebble100-24              993.8n ± 2%   747.1n ± 1%  -24.82% (p=0.000 n=11)
Read/LookasideDistPebble1000-24             110.2µ ± 1%   108.6µ ± 0%   -1.39% (p=0.000 n=11)
Read/LookasideDistPebble10000-24            150.1µ ± 1%   149.4µ ± 2%        ~ (p=0.519 n=11)
GetSingle/LookasideDistPebble10/AC-24       102.9µ ± 1%   102.7µ ± 1%        ~ (p=0.438 n=11)
GetSingle/LookasideDistPebble10/CAS-24      874.4n ± 3%   682.4n ± 2%  -21.96% (p=0.000 n=11)
GetSingle/LookasideDistPebble100/AC-24      105.8µ ± 1%   104.9µ ± 1%   -0.81% (p=0.023 n=11)
GetSingle/LookasideDistPebble100/CAS-24     908.9n ± 3%   722.5n ± 2%  -20.51% (p=0.000 n=11)
GetSingle/LookasideDistPebble1000/AC-24     107.7µ ± 4%   106.6µ ± 1%   -0.99% (p=0.007 n=11)
GetSingle/LookasideDistPebble1000/CAS-24    111.8µ ± 2%   109.0µ ± 1%   -2.52% (p=0.000 n=11)
GetSingle/LookasideDistPebble10000/AC-24    136.3µ ± 1%   135.9µ ± 1%        ~ (p=0.171 n=11)
GetSingle/LookasideDistPebble10000/CAS-24   139.7µ ± 1%   137.6µ ± 1%   -1.46% (p=0.000 n=11)
FindMissing/LookasideDistPebble10-24        356.1µ ± 4%   341.5µ ± 5%   -4.09% (p=0.034 n=11)
FindMissing/LookasideDistPebble100-24       486.6µ ± 3%   461.8µ ± 2%   -5.11% (p=0.000 n=11)
FindMissing/LookasideDistPebble1000-24      492.2µ ± 2%   472.8µ ± 2%   -3.94% (p=0.000 n=11)
FindMissing/LookasideDistPebble10000-24     476.8µ ± 2%   468.6µ ± 1%   -1.71% (p=0.002 n=11)
geomean                                     49.06µ        45.34µ        -7.58%

                                          │     base     │              findmissing2               │
                                          │     B/s      │      B/s       vs base                  │
Read/LookasideDistPebble10-24               11.22Mi ± 2%    14.62Mi ± 2%  +30.36% (p=0.000 n=11)
Read/LookasideDistPebble100-24              95.96Mi ± 4%   127.65Mi ± 1%  +33.03% (p=0.000 n=11)
Read/LookasideDistPebble1000-24             8.659Mi ± 1%    8.774Mi ± 1%   +1.32% (p=0.000 n=11)
Read/LookasideDistPebble10000-24            63.52Mi ± 1%    63.84Mi ± 3%        ~ (p=0.519 n=11)
GetSingle/LookasideDistPebble10/AC-24       97.66Ki ± 0%    97.66Ki ± 0%        ~ (p=1.000 n=11) ¹
GetSingle/LookasideDistPebble10/CAS-24      10.91Mi ± 2%    13.97Mi ± 2%  +28.06% (p=0.000 n=11)
GetSingle/LookasideDistPebble100/AC-24      927.7Ki ± 1%    927.7Ki ± 1%    0.00% (p=0.025 n=11)
GetSingle/LookasideDistPebble100/CAS-24     104.9Mi ± 3%    132.0Mi ± 2%  +25.80% (p=0.000 n=11)
GetSingle/LookasideDistPebble1000/AC-24     8.850Mi ± 6%    8.945Mi ± 1%   +1.08% (p=0.005 n=11)
GetSingle/LookasideDistPebble1000/CAS-24    8.526Mi ± 2%    8.755Mi ± 1%   +2.68% (p=0.000 n=11)
GetSingle/LookasideDistPebble10000/AC-24    69.99Mi ± 1%    70.18Mi ± 0%        ~ (p=0.176 n=11)
GetSingle/LookasideDistPebble10000/CAS-24   68.28Mi ± 3%    69.29Mi ± 1%   +1.48% (p=0.000 n=11)
geomean                                     12.89Mi         14.13Mi        +9.60%
¹ all samples are equal

                                          │     base     │             findmissing2             │
                                          │     B/op     │     B/op      vs base                │
Read/LookasideDistPebble10-24                 903.0 ± 0%     559.0 ± 0%  -38.10% (p=0.000 n=11)
Read/LookasideDistPebble100-24               1152.0 ± 1%     752.0 ± 1%  -34.72% (p=0.000 n=11)
Read/LookasideDistPebble1000-24             35.81Ki ± 0%   35.09Ki ± 0%   -2.00% (p=0.000 n=11)
Read/LookasideDistPebble10000-24            70.52Ki ± 1%   69.91Ki ± 1%        ~ (p=0.151 n=11)
GetSingle/LookasideDistPebble10/AC-24       29.52Ki ± 0%   29.19Ki ± 0%   -1.12% (p=0.000 n=11)
GetSingle/LookasideDistPebble10/CAS-24        937.0 ± 0%     601.0 ± 0%  -35.86% (p=0.000 n=11)
GetSingle/LookasideDistPebble100/AC-24      29.93Ki ± 0%   29.58Ki ± 0%   -1.17% (p=0.000 n=11)
GetSingle/LookasideDistPebble100/CAS-24      1037.0 ± 0%     701.0 ± 0%  -32.40% (p=0.000 n=11)
GetSingle/LookasideDistPebble1000/AC-24     33.83Ki ± 0%   33.50Ki ± 0%   -0.97% (p=0.000 n=11)
GetSingle/LookasideDistPebble1000/CAS-24    34.85Ki ± 0%   34.20Ki ± 0%   -1.87% (p=0.000 n=11)
GetSingle/LookasideDistPebble10000/AC-24    50.31Ki ± 0%   49.98Ki ± 0%   -0.67% (p=0.000 n=11)
GetSingle/LookasideDistPebble10000/CAS-24   60.45Ki ± 0%   59.78Ki ± 0%   -1.11% (p=0.000 n=11)
FindMissing/LookasideDistPebble10-24        265.2Ki ± 8%   243.2Ki ± 8%   -8.28% (p=0.005 n=11)
FindMissing/LookasideDistPebble100-24       378.4Ki ± 0%   345.4Ki ± 0%   -8.71% (p=0.000 n=11)
FindMissing/LookasideDistPebble1000-24      409.5Ki ± 0%   376.2Ki ± 0%   -8.11% (p=0.000 n=11)
FindMissing/LookasideDistPebble10000-24     379.6Ki ± 0%   346.6Ki ± 0%   -8.67% (p=0.000 n=11)
geomean                                     27.62Ki        24.08Ki       -12.81%

                                          │    base     │            findmissing2             │
                                          │  allocs/op  │  allocs/op   vs base                │
Read/LookasideDistPebble10-24                14.00 ± 0%    12.00 ± 0%  -14.29% (p=0.000 n=11)
Read/LookasideDistPebble100-24               15.00 ± 0%    13.00 ± 0%  -13.33% (p=0.000 n=11)
Read/LookasideDistPebble1000-24              476.0 ± 0%    472.0 ± 0%   -0.84% (p=0.000 n=11)
Read/LookasideDistPebble10000-24             472.0 ± 0%    468.0 ± 0%   -0.85% (p=0.000 n=11)
GetSingle/LookasideDistPebble10/AC-24        471.0 ± 0%    469.0 ± 0%   -0.42% (p=0.000 n=11)
GetSingle/LookasideDistPebble10/CAS-24       16.00 ± 0%    14.00 ± 0%  -12.50% (p=0.000 n=11)
GetSingle/LookasideDistPebble100/AC-24       471.0 ± 0%    469.0 ± 0%   -0.42% (p=0.000 n=11)
GetSingle/LookasideDistPebble100/CAS-24      17.00 ± 0%    15.00 ± 0%  -11.76% (p=0.000 n=11)
GetSingle/LookasideDistPebble1000/AC-24      471.0 ± 0%    469.0 ± 0%   -0.42% (p=0.000 n=11)
GetSingle/LookasideDistPebble1000/CAS-24     478.0 ± 0%    474.0 ± 0%   -0.84% (p=0.000 n=11)
GetSingle/LookasideDistPebble10000/AC-24     467.0 ± 0%    465.0 ± 0%   -0.43% (p=0.000 n=11)
GetSingle/LookasideDistPebble10000/CAS-24    474.0 ± 0%    470.0 ± 0%   -0.84% (p=0.000 n=11)
FindMissing/LookasideDistPebble10-24        4.164k ± 7%   4.032k ± 7%        ~ (p=0.171 n=11)
FindMissing/LookasideDistPebble100-24       6.137k ± 0%   5.937k ± 0%   -3.26% (p=0.000 n=11)
FindMissing/LookasideDistPebble1000-24      6.137k ± 0%   5.937k ± 0%   -3.26% (p=0.000 n=11)
FindMissing/LookasideDistPebble10000-24     6.137k ± 0%   5.937k ± 0%   -3.26% (p=0.000 n=11)
geomean                                      372.4         355.6        -4.51%
```